TypeScript Refactor
This was inevitable, given that variable types are half of the point.
