+++ .editorconfig
@@ -10,5 +10,5 @@ indent_size = 2
 insert_final_newline = true
 trim_trailing_whitespace = true
 
-[*.{md, diff,patch}]
+[*.{md,diff,patch}]
 trim_trailing_whitespace = false
