+++ .eslintignore
@@ -1,0 +1,3 @@
+lib
+coverage
+examples/basic/index.mjs
