+++ .github/workflows/ci.yml
@@ -17,6 +17,19 @@ jobs:
         env:
           CI: true
 
+  build:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - name: Setup Node
+        uses: actions/setup-node@v1
+        with:
+          node-version: 14.x
+      - run: npm install
+      - run: npm run build
+        env:
+          CI: true
+
   test:
     runs-on: ${{ matrix.os }}
     strategy:
