+++ .gitignore
@@ -1,3 +1,5 @@
 node_modules
 coverage
+lib
+
 RELEASE.md
