+++ .npmignore
@@ -1,3 +1,4 @@
+src
 examples
 tests
 coverage
@@ -6,5 +7,8 @@ coverage
 .editorconfig
 .gitattributes
 .github
+.vscode
+jest.config.json
+tsconfig.json
 RELEASE.md.hbs
 RELEASE.md
