+++ .vscode/settings.json
@@ -1,0 +1,7 @@
+{
+  "typescript.tsdk": "node_modules/typescript/lib",
+  "search.exclude": {
+    "/lib": true,
+    "/coverage": true
+  }
+}
