+++ examples/basic/index.cjs
