+++ jest.config.json
@@ -1,0 +1,4 @@
+{
+  "preset": "ts-jest",
+  "testEnvironment": "node"
+}
