+++ src/index.ts
@@ -1,0 +1,6 @@
+import load from './load';
+import type from './type';
+import { parse, parseFile } from './parse';
+
+export { load, type, parse, parseFile };
+export default { load, type, parse, parseFile };
